Add garbage collection; additional checks for easter egg
Add garbage collection, the lack of which caused the bot to stop before. I have tested with
garbace collection on for a few days, and it seems to work flawlessly with it.

Also do some additional checks for the emoji easter egg – check whether a character is a word
character (i.e. regex \w), and check whether it is in ASCII. If it is either, use the normal
separator instead of the "emoji". This should fix #5.