Add selection of lemma with @langcode; display language name for lexemes
Let a user specify a specific language variety when linking by
adding @langcode to the Lexeme ID. Example: L71370@ms-arab will only
show the Arab-script lemma from L71370. The name of the lexeme's language
will also be shown after the language code, with a link to its item.

This fixes #23.